// schema.graphql/API.ts
export type CreateBusinessMutationVariables = {
  condition?: ModelBusinessConditionInput | null,
  input: CreateBusinessInput,
};

export type CreateBusinessMutation = {
  createBusiness?:  {
    __typename: "Business",
    address: string,
    billingType?: BusinessBillingType | null,
    businessType?: BusinessBusinessType | null,
    city: string,
    cpmRate?: number | null,
    createdAt: string,
    currentSpend?: number | null,
    description?: string | null,
    email: string,
    id: string,
    isActive?: boolean | null,
    isVerified?: boolean | null,
    latitude?: number | null,
    logoKey?: string | null,
    logoUrl?: string | null,
    longitude?: number | null,
    monthlyBudget?: number | null,
    name: string,
    owner: string,
    phone?: string | null,
    rejectionReason?: string | null,
    rewards?:  {
      __typename: "ModelRewardConnection",
      nextToken?: string | null,
    } | null,
    serviceRadius?: number | null,
    state: string,
    updatedAt: string,
    verificationStatus?: BusinessVerificationStatus | null,
    website?: string | null,
    zipCode: string,
  } | null,
};

export type CreateBusinessAnalyticsMutationVariables = {
  condition?: ModelBusinessAnalyticsConditionInput | null,
  input: CreateBusinessAnalyticsInput,
};

export type CreateBusinessAnalyticsMutation = {
  createBusinessAnalytics?:  {
    __typename: "BusinessAnalytics",
    avgChildAge?: number | null,
    businessId: string,
    clicks?: number | null,
    createdAt: string,
    date: string,
    id: string,
    impressions?: number | null,
    owner?: string | null,
    redemptions?: number | null,
    revenue?: number | null,
    spend?: number | null,
    topRewardId?: string | null,
    updatedAt: string,
  } | null,
};

export type CreateRewardMutationVariables = {
  condition?: ModelRewardConditionInput | null,
  input: CreateRewardInput,
};

export type CreateRewardMutation = {
  createReward?:  {
    __typename: "Reward",
    availableRadius?: number | null,
    business?:  {
      __typename: "Business",
      address: string,
      billingType?: BusinessBillingType | null,
      businessType?: BusinessBusinessType | null,
      city: string,
      cpmRate?: number | null,
      createdAt: string,
      currentSpend?: number | null,
      description?: string | null,
      email: string,
      id: string,
      isActive?: boolean | null,
      isVerified?: boolean | null,
      latitude?: number | null,
      logoKey?: string | null,
      logoUrl?: string | null,
      longitude?: number | null,
      monthlyBudget?: number | null,
      name: string,
      owner: string,
      phone?: string | null,
      rejectionReason?: string | null,
      serviceRadius?: number | null,
      state: string,
      updatedAt: string,
      verificationStatus?: BusinessVerificationStatus | null,
      website?: string | null,
      zipCode: string,
    } | null,
    businessId: string,
    category?: RewardCategory | null,
    clicks?: number | null,
    createdAt: string,
    currentRedemptions?: number | null,
    dailyLimit?: number | null,
    daysOfWeek?: Array< string | null > | null,
    description: string,
    endDate?: string | null,
    id: string,
    imageKey?: string | null,
    imageUrl?: string | null,
    impressions?: number | null,
    isActive?: boolean | null,
    owner?: string | null,
    pointCost: number,
    redemptionInstructions?: string | null,
    redemptionRecords?:  {
      __typename: "ModelRewardRedemptionConnection",
      nextToken?: string | null,
    } | null,
    redemptionType?: RewardRedemptionType | null,
    redemptions?: number | null,
    retailValue?: number | null,
    rewardType?: RewardRewardType | null,
    startDate: string,
    status?: RewardStatus | null,
    targetAgeRanges?: Array< string | null > | null,
    targetFamilySizes?: Array< number | null > | null,
    termsAndConditions?: string | null,
    title: string,
    totalAvailable?: number | null,
    updatedAt: string,
  } | null,
};

export type CreateRewardRedemptionMutationVariables = {
  condition?: ModelRewardRedemptionConditionInput | null,
  input: CreateRewardRedemptionInput,
};

export type CreateRewardRedemptionMutation = {
  createRewardRedemption?:  {
    __typename: "RewardRedemption",
    businessId?: string | null,
    businessRevenue?: number | null,
    childAge?: number | null,
    childId: string,
    createdAt: string,
    expiresAt?: string | null,
    familyId: string,
    fulfilledAt?: string | null,
    id: string,
    notes?: string | null,
    owner?: string | null,
    pointsSpent: number,
    redemptionCity?: string | null,
    redemptionCode?: string | null,
    redemptionDate?: string | null,
    redemptionStatus?: RewardRedemptionRedemptionStatus | null,
    redemptionZip?: string | null,
    reward?:  {
      __typename: "Reward",
      availableRadius?: number | null,
      businessId: string,
      category?: RewardCategory | null,
      clicks?: number | null,
      createdAt: string,
      currentRedemptions?: number | null,
      dailyLimit?: number | null,
      daysOfWeek?: Array< string | null > | null,
      description: string,
      endDate?: string | null,
      id: string,
      imageKey?: string | null,
      imageUrl?: string | null,
      impressions?: number | null,
      isActive?: boolean | null,
      owner?: string | null,
      pointCost: number,
      redemptionInstructions?: string | null,
      redemptionType?: RewardRedemptionType | null,
      redemptions?: number | null,
      retailValue?: number | null,
      rewardType?: RewardRewardType | null,
      startDate: string,
      status?: RewardStatus | null,
      targetAgeRanges?: Array< string | null > | null,
      targetFamilySizes?: Array< number | null > | null,
      termsAndConditions?: string | null,
      title: string,
      totalAvailable?: number | null,
      updatedAt: string,
    } | null,
    rewardId: string,
    updatedAt: string,
  } | null,
};

export type DeleteBusinessMutationVariables = {
  condition?: ModelBusinessConditionInput | null,
  input: DeleteBusinessInput,
};

export type DeleteBusinessMutation = {
  deleteBusiness?:  {
    __typename: "Business",
    address: string,
    billingType?: BusinessBillingType | null,
    businessType?: BusinessBusinessType | null,
    city: string,
    cpmRate?: number | null,
    createdAt: string,
    currentSpend?: number | null,
    description?: string | null,
    email: string,
    id: string,
    isActive?: boolean | null,
    isVerified?: boolean | null,
    latitude?: number | null,
    logoKey?: string | null,
    logoUrl?: string | null,
    longitude?: number | null,
    monthlyBudget?: number | null,
    name: string,
    owner: string,
    phone?: string | null,
    rejectionReason?: string | null,
    rewards?:  {
      __typename: "ModelRewardConnection",
      nextToken?: string | null,
    } | null,
    serviceRadius?: number | null,
    state: string,
    updatedAt: string,
    verificationStatus?: BusinessVerificationStatus | null,
    website?: string | null,
    zipCode: string,
  } | null,
};

export type DeleteBusinessAnalyticsMutationVariables = {
  condition?: ModelBusinessAnalyticsConditionInput | null,
  input: DeleteBusinessAnalyticsInput,
};

export type DeleteBusinessAnalyticsMutation = {
  deleteBusinessAnalytics?:  {
    __typename: "BusinessAnalytics",
    avgChildAge?: number | null,
    businessId: string,
    clicks?: number | null,
    createdAt: string,
    date: string,
    id: string,
    impressions?: number | null,
    owner?: string | null,
    redemptions?: number | null,
    revenue?: number | null,
    spend?: number | null,
    topRewardId?: string | null,
    updatedAt: string,
  } | null,
};

export type DeleteRewardMutationVariables = {
  condition?: ModelRewardConditionInput | null,
  input: DeleteRewardInput,
};

export type DeleteRewardMutation = {
  deleteReward?:  {
    __typename: "Reward",
    availableRadius?: number | null,
    business?:  {
      __typename: "Business",
      address: string,
      billingType?: BusinessBillingType | null,
      businessType?: BusinessBusinessType | null,
      city: string,
      cpmRate?: number | null,
      createdAt: string,
      currentSpend?: number | null,
      description?: string | null,
      email: string,
      id: string,
      isActive?: boolean | null,
      isVerified?: boolean | null,
      latitude?: number | null,
      logoKey?: string | null,
      logoUrl?: string | null,
      longitude?: number | null,
      monthlyBudget?: number | null,
      name: string,
      owner: string,
      phone?: string | null,
      rejectionReason?: string | null,
      serviceRadius?: number | null,
      state: string,
      updatedAt: string,
      verificationStatus?: BusinessVerificationStatus | null,
      website?: string | null,
      zipCode: string,
    } | null,
    businessId: string,
    category?: RewardCategory | null,
    clicks?: number | null,
    createdAt: string,
    currentRedemptions?: number | null,
    dailyLimit?: number | null,
    daysOfWeek?: Array< string | null > | null,
    description: string,
    endDate?: string | null,
    id: string,
    imageKey?: string | null,
    imageUrl?: string | null,
    impressions?: number | null,
    isActive?: boolean | null,
    owner?: string | null,
    pointCost: number,
    redemptionInstructions?: string | null,
    redemptionRecords?:  {
      __typename: "ModelRewardRedemptionConnection",
      nextToken?: string | null,
    } | null,
    redemptionType?: RewardRedemptionType | null,
    redemptions?: number | null,
    retailValue?: number | null,
    rewardType?: RewardRewardType | null,
    startDate: string,
    status?: RewardStatus | null,
    targetAgeRanges?: Array< string | null > | null,
    targetFamilySizes?: Array< number | null > | null,
    termsAndConditions?: string | null,
    title: string,
    totalAvailable?: number | null,
    updatedAt: string,
  } | null,
};

export type DeleteRewardRedemptionMutationVariables = {
  condition?: ModelRewardRedemptionConditionInput | null,
  input: DeleteRewardRedemptionInput,
};

export type DeleteRewardRedemptionMutation = {
  deleteRewardRedemption?:  {
    __typename: "RewardRedemption",
    businessId?: string | null,
    businessRevenue?: number | null,
    childAge?: number | null,
    childId: string,
    createdAt: string,
    expiresAt?: string | null,
    familyId: string,
    fulfilledAt?: string | null,
    id: string,
    notes?: string | null,
    owner?: string | null,
    pointsSpent: number,
    redemptionCity?: string | null,
    redemptionCode?: string | null,
    redemptionDate?: string | null,
    redemptionStatus?: RewardRedemptionRedemptionStatus | null,
    redemptionZip?: string | null,
    reward?:  {
      __typename: "Reward",
      availableRadius?: number | null,
      businessId: string,
      category?: RewardCategory | null,
      clicks?: number | null,
      createdAt: string,
      currentRedemptions?: number | null,
      dailyLimit?: number | null,
      daysOfWeek?: Array< string | null > | null,
      description: string,
      endDate?: string | null,
      id: string,
      imageKey?: string | null,
      imageUrl?: string | null,
      impressions?: number | null,
      isActive?: boolean | null,
      owner?: string | null,
      pointCost: number,
      redemptionInstructions?: string | null,
      redemptionType?: RewardRedemptionType | null,
      redemptions?: number | null,
      retailValue?: number | null,
      rewardType?: RewardRewardType | null,
      startDate: string,
      status?: RewardStatus | null,
      targetAgeRanges?: Array< string | null > | null,
      targetFamilySizes?: Array< number | null > | null,
      termsAndConditions?: string | null,
      title: string,
      totalAvailable?: number | null,
      updatedAt: string,
    } | null,
    rewardId: string,
    updatedAt: string,
  } | null,
};